optscale-deploy/requirements.txt
ansible==9.0.1
pyyaml==6.0.1
kubernetes==10.0.1
python-on-whales==0.72.0
